refactor client.py, simple entrypoint and support new process without side effects
